requirements.txt
# App & Server Framework
fastapi==0.115.0
uvicorn==0.30.6

# Database & Config
sqlalchemy==2.0.35
pyodbc
python-dotenv==1.0.1
pydantic-settings==2.5.2
pydantic

# LlamaIndex RAG
llama-index
llama-index-llms-gemini
llama-index-embeddings-gemini

# Document Parsers
pypdf
python-docx==1.1.2

# Frontend & Uploads
streamlit==1.38.0
python-multipart==0.0.9